Rewards based on difficulty
Introduced BASE_WORK_UNIT = 100 which gets now multiplied with the difficulty multiplier of a work unit.
It's set to 100 so that a multiplier of 1.5x still has a good effect (results in reward to be 150).

Since this is a server side change only and all payouts are percentage based on the difference between total WU and payed WU and the totals are calculated from all clients aggregated no code changes are needed in the payout and or stats logic.

Since no datatypes have changed (WU is still an int) no client update and/or db migration is needed.